Add support for compilation from the command line; retire TUI
This adds a way to compile programs from the command line. The command
line parsing is very rigid for now, but compilation should be working.
To limit somewhat the amount of code included in the binaries,
currently the eval support is disabled. There's a lot of room for
improvements.

The TUI code was also removed. It will be better to keep it out of the
regular Loko build and provide it as normal application code.

Makefiles are used instead of scripts.

Signed-off-by: Göran Weinholt <goran@weinholt.se>