fibers: Store timers in heaps and add cleanup of channels
Timers are now stored in heaps (from the pfds package), which should
be more efficient.

When choice-operation completes, only one block function is called, so
other operations are never cleaned up. This commit adds some cleanup
code for channels and conditions.

Timer operations should probably be cleaned up as well, but that isn't
done in this commit. Each choice that has a timer that didn't sync
leaves an timer in the heap, which can get unnecessarily expensive.

Signed-off-by: Göran Weinholt <goran@weinholt.se>