fibers: Rewrite the UART driver to use fibers, various IRQ rework
The UART driver used by pid 1 is now based on fibers, which means that
I/O in fibers is non-blocking even when running on bare hardware. (The
UART driver uses interrupt-operation to wait for an interrupt, but
this mechanism is not fully public yet).

The debug output from pid 0 is now sent to the debug console so that
it will not interfere with the UART driver. The debug console is a
device in QEMU accessible via -debugcon vc.

The generic interrupt handler checks if CPL=0 and then does not
preempt the process. This is mostly a sanity thing.

Some more syscalls are now using vectors. However the interface should
be made consistent and it should enable the scheduler to always pass a
message to the process.

Signed-off-by: Göran Weinholt <goran@weinholt.se>