Replace allocate with dma-allocate
This removes the allocate procedure and replaces it with dma-allocate.
The latter is intended to be more specific to the needs of DMA
operations. It also returns just one address. It should be possible to
keep all DMA:able addresses identity mapped; then no software
virt->phys translation is needed.

Signed-off-by: Göran Weinholt <goran@weinholt.se>