Merge loko-drivers, loko-posix, rename libs to runtime
I take back my previous idea of splitting off the drivers, POSIX API,
etc into separate projects. It will be confusing and cause unnecessary
overhead.

The libs/ directory is renamed to runtime/, to better reflect what it
is. This means they are libraries not directly meant for end users.
End user libraries are installed by the makefile and also includes
those built-in libraries listed in compiler/expander.sls that don't
have a $ in the name.

Signed-off-by: Göran Weinholt <goran@weinholt.se>