Fix miscompilation of bytevector-[su]{16,32}-ref
This fixes a miscompilation of the 16-bit and 32-bit bytevector
procedures. If the register allocator assigned a high register (r8 and
up) to the bytevector, then the generated instruction used the "ah"
register, which is not possible. Unfortunately the assembler doesn't
check for this particular error. The actual result is that the
instruction writes to "spl", the lower eight bits of the stack
register.

The output from the code generator (after optimization):

  (mov ah (mem8+ rcx r9 1 8 8 -5))
     ;sets: (rax) uses: (r9 rcx rax) deads: (r9 rcx rax)

When disassembled:

  210A57: 428A64090C      (mov spl (mem8+ rcx #xC (* r9 #x1)))

Fixes: #19
Signed-off-by: Göran Weinholt <goran@weinholt.se>