compiler: Add the compiler option 'freestanding
This adds 'freestanding to the options (-ffreestanding on the command
line), which disables all built-in Scheme libraries. The assembler
libraries provided by the target are still built, such as the GC and
the debug printer.

This flag is useful mostly to test code generators and the low-level
optimizer. It's also useful when developing new code generators
because only the code contained in the compiled program will be passed
to the code generator, making it easier to develop step by step.

Fixes: #24
Signed-off-by: Göran Weinholt <goran@weinholt.se>