arch: Split asm code to reduce hard dependencies on amd64
This splits the (loko asm) library to different libraries for each
supported target. The new (loko asm) library has a registry of
supported architectures. This makes it easier to add new targets and
not import the amd64 libraries.

As a related change, it is now also possible to build for a selected
target kernel. The command line argument -ftarget= selects the target,
which is currently one of linux, pc or pc+linux. The new just-hello
sample shows how to create a 16K "hello world" with the freestanding
compile mode.

The "loko" target is renamed to "pc" because "loko" as such does not
describe the environment where the program will be running. There are
more AMD64 bare metal hardware types that could be supported.

Signed-off-by: Göran Weinholt <goran@weinholt.se>