lib posix: Update to SRFI 170 draft #7
Removed the error procedures and the errno aliases. A few procedures
lost support for taking ports. Updated open-directory to take an
optional dot-files? argument.

Implemented create-temp-file and call-with-temporary-filename.

Worked around a problem in the with-/without- procedures due to
dynamic-wind not composing neatly with fibers and ports when the
handlers modify the ports (quite a specific case, but there it is).

Signed-off-by: Göran Weinholt <goran@weinholt.se>