Fix booting on systems with more than 4GB of RAM
This fixes a boot issue that was introduced with the buddy allocator.
The buddy allocators write to the area they cover, so the memory needs
to be mapped first. Only the first 4GB of RAM has mappings in the
initial page table.

The fix is to make buddy allocators for the areas that are already
mapped, then use those to allocate pages for the high areas, and
finally create buddy allocators for the the high areas.

Signed-off-by: Göran Weinholt <goran@weinholt.se>