srfi: Add and install (srfi :170 posix), MIT-licensed
The SRFI 170 implementation is now available under the standard names
(srfi :170 posix) and (srfi :170). The "posix" part is mostly a guess;
I assume most implementors will use that name.

The srfi/ directory will be under the MIT license, the normal one for
SRFIs, in case someone in the wider Scheme community wants to use it.

Signed-off-by: Göran Weinholt <goran@weinholt.se>