Makefile: add a faster self compilation target
This adds a faster self compilation target and renames the old one.
The new one uses the -feval flag to enable eval and simply builds Loko
as if it were any other Scheme program.

The new target's build time is about 2.5 times that of the loko
target, which bootstraps with Chez Scheme. The old target was ~5x.

Signed-off-by: Göran Weinholt <goran@weinholt.se>