drivers ps2: PS/2 drivers for keyboards (partially working) and mice
This adds drivers for the i8042 controller in the PC (and some other
machines), which handles the PS/2 legacy ports. These are used in
emulators and sometimes in laptops.

PS/2 drivers for keyboards and mice are also added in this commit. The
mouse support should be fairly complete, but the keyboards are not
being decoded completely yet. For lack of anything better, I will be
converting the scancodes to USB HID boot keyboard codes.

Signed-off-by: Göran Weinholt <goran@weinholt.se>