arch amd64: Fix logic for the 8259/PIC's Special Mask Mode
The way that Loko handled the PIC meant that it was still working in
nested mode. This was not intended, since it would mean a slow or
misbehaving driver on a high priority IRQ would block a lower priority
IRQ. ("Priority" is just old school terminology here and just means a
lower IRQ number).

Special Mask Mode relies on the kernel masking the IRQs it receives.
It then sends a Specific End Of Interrupt to acknowledge that IRQ.
When the driver has handled the IRQ we simply unmask it. The names of
the library procedures keep their names and work as expected, but the
mechanics behind them are changed to get the correct semantics.

Thanks to Christian Häggström for drawing my attention to this
requirement of Special Mask Mode.

Signed-off-by: Göran Weinholt <goran@weinholt.se>