arch amd64: Do not enable IRQ 0
While testing on another PC, I noticed that IRQ 0 would be picked up
in the PIC but not delivered to the CPU. This is likely because IRQ 0
is edge triggered. The lost delivery of IRQ 0 meant that every other
IRQ on the system was blocked. The intention was to clear IRQ 0 in the
IRR, but it doesn't appear to cause any harm if it is set there.

It might be necessary to actually poll the PICs for missed edge
triggered interrupts and deliver them manually. Setting the DEBUG
variable to #t is a good way to see if that is what's blocking the
system.

Signed-off-by: Göran Weinholt <goran@weinholt.se>