drivers pci: New library for reading PCI expansion ROMS
This adds the (loko drivers pci roms) library that lets you read and
parse expansion ROMs. These are firmware blobs built in to the
hardware that provide functionality to the BIOS, EFI or equivalent.

It could be improved to handle the case where the vendor didn't add a
BAR for this purpose, but it's probably good enough for now.

Signed-off-by: Göran Weinholt <goran@weinholt.se>