runtime arithmetic: Less wrong int-inexact
Previous int-inexact added together all the digits of the bignum, but
this gives very wrong results, very fast. It now takes the most
significant bits, converts that to a flonum and then gets the right
exponent by iterative doubling. It needs to be tested for corner
cases, but it's much better than it was before.

Signed-off-by: Göran Weinholt <goran@weinholt.se>